requirements.txt
# 核心依賴
PyYAML>=6.0
ruamel.yaml>=0.18.0
tiktoken>=0.5.1
requests>=2.31.0
click>=8.1.7
python-dotenv>=1.0.0
tqdm>=4.66.1

# 開發依賴
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
